requirements.txt
# Workshop Requirements
# Install with: pip install -r requirements.txt

# Core dependencies
openai>=1.12.0
python-dotenv>=1.0.0

# Day 3 - FastAPI (for production API)
fastapi>=0.104.0
uvicorn>=0.24.0
pydantic>=2.0.0

# Day 3 - Optional (for testing API)
requests>=2.31.0

